don't add the faulty cc# mirror to the round-robin cc.php.net record.
this means that if all of the cc# mirrors for a cc are faulty the
cc.php.net record will be missing from the zone.

we had a discussion on irc with bjori and talked about whether or not
should we fall back using the www.php.net mirror if a cc is down, but
we decided to go with the simpler solution for now, we can tweak it
later.