Use OhDear as status-page for php.net
Currently calling status.php.net will redirect the caller to php.net
which does not make too much sense. As OhDear has offered to sponsor
status-monitoring for php.net this change provides a DNS-record for
displaying a decent status page that we do not need to take care of
apart from configuring.