Write log into repo specific file
That way the output of the git reset command will be output into one
file per repo instead of only containing the output of the last
assertion. And it avoids a cleanup at the begining of the script.

An alternative would be using `logger` instead but that would then
preserve the output completely and lead to an increased size of the
syslog. Not sure that makes sense.